Moved most io code into a new package
The io package was added as the location for format specific code. Currently there is only a BinaryFormatter class, implementing the behavior of the previous code. Currently there is a bug where it does not parse correctly, but as it runs and I've been working on this for a while, I decided to push it anyways.

The object package has been updated with several changes, most of which are focused on decoupling the value objects from the record format of the NRBF specification.

The record package has been removed entirely, as the reason for it was to implement the read/write functions for each record type which has been moved to the dotnet.io.binary.BinaryFormatter class.

The primitives package has been stripped of the write(), read() and __bytes__() methods, as those implemented details which have been moved to the io package.