Removed object_id from instances and started work on write methods
The object_id of the Instance class was removed, in favor of a plain
reference to the DataStore on which the Instance was defined. The object
id was not used in any meaningful capacity, as doing so would have
required writing a garbage collector of sorts.

Some basic write methods have been added as well. The write methods
currently can not be used to fully write out any valid Instance; there
are several missing methods at the moment.