Edits to the Library class and classes which use it
The Library class now properly parses specification strings according to
MS-NRTP 2.2.1.3 LibraryName. Further, the LibraryID has been removed
from the Library class. Classes which use the Library class can use the
hash() function on the library to obtain an integral key for Libraries.
The DataStore now stores the libraries as a mapping between this hash
key and the Library itself, to allow looking up the Library when freshly
parsed from the stream such that there is only one Library of a given
name and version at any given point in time.